+++ src/components/TourButton/TourButton.module.css
@@ -1,4 +1,4 @@
-.tourButton {
+/* .tourButton {
     position: fixed;
     bottom: 10px;
     right: 5px;
@@ -14,4 +14,4 @@
     width: 40px;
     height: 40px;
     fill: #555;
-}
+} */
