+++ src/pages/TrackerPage/TrackerPage.module.css
@@ -8,6 +8,23 @@
   margin: 0 auto;
 }
 
+.tourButton {
+  position: fixed;
+  bottom: 10px;
+  right: 5px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  z-index: 1000;
+  background-color: transparent;
+  /* border-style: none; */
+}
+
+.question {
+  width: 40px;
+  height: 40px;
+  fill: #555;
+}
 
 .shake {
   animation: shake 1s ease-in-out;
